requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-multipart>=0.0.5
PyJWT>=2.1.0
qrcode[pil]>=7.3.1
Pillow>=8.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=0.19.0
